refactor: move c9 related activation code to shared
